refactor: AuctionStatus Enum 클래스 분리 (#CHZZ-80)
