+++ src/test/java/org/chzz/market/domain/bid/service/BidServiceTest.java
@@ -16,7 +16,7 @@ import static org.mockito.Mockito.when;
 import java.time.LocalDateTime;
 import java.util.Optional;
 import org.chzz.market.domain.auction.entity.Auction;
-import org.chzz.market.domain.auction.entity.Auction.AuctionStatus;
+import org.chzz.market.domain.auction.enums.AuctionStatus;
 import org.chzz.market.domain.auction.error.AuctionException;
 import org.chzz.market.domain.auction.service.AuctionService;
 import org.chzz.market.domain.bid.dto.BidCreateRequest;
